requirements.txt
pandas
streamlit
requests
plotly
python-dotenv